+++ web-frontend/modules/database/fieldTypes.js
@@ -215,7 +215,7 @@ export class FieldType extends Registerable {
   /**
    * Indicates whether or not it is possible to sort in a view.
    */
-  getCanSortInView() {
+  getCanSortInView(field) {
     return true
   }
 
@@ -230,7 +230,6 @@ export class FieldType extends Registerable {
     super(...args)
     this.type = this.getType()
     this.iconClass = this.getIconClass()
-    this.canSortInView = this.getCanSortInView()
     this.canBePrimaryField = this.getCanBePrimaryField()
     this.isReadOnly = this.getIsReadOnly()
 
@@ -263,7 +262,6 @@ export class FieldType extends Registerable {
       type: this.type,
       iconClass: this.iconClass,
       name: this.getName(),
-      canSortInView: this.canSortInView,
       isReadOnly: this.isReadOnly,
     }
   }
@@ -672,7 +670,7 @@ export class LinkRowFieldType extends FieldType {
     return []
   }
 
-  getCanSortInView() {
+  getCanSortInView(field) {
     return false
   }
 
@@ -1586,7 +1584,7 @@ export class FileFieldType extends FieldType {
     return []
   }
 
-  getCanSortInView() {
+  getCanSortInView(field) {
     return false
   }
 
@@ -2007,7 +2005,7 @@ export class FormulaFieldType extends FieldType {
   static compatibleWithFormulaTypes(...formulaTypeStrings) {
     return (field) => {
       return (
-        (field.type === this.getType() || field.type === 'lookup') &&
+        field.type === this.getType() &&
         formulaTypeStrings.includes(field.formula_type)
       )
     }
@@ -2133,6 +2131,11 @@ export class FormulaFieldType extends FieldType {
 
   canBeReferencedByFormulaField() {
     return true
+  }
+
+  getCanSortInView(field) {
+    const subType = this.app.$registry.get('formula_type', field.formula_type)
+    return subType.getCanSortInView()
   }
 }